mod: refactor LineSearch initializer
I came back to LineSearch class after writing it > 3 years earlier and
it was hard for me to understand how to instantiate a LineSearch object.
Many arguments were optional and typing and docstring was missing.

By the virtue of functools.singledispatchmethod a LineSearch-object is
now easily created from a Geometry and a search direction, OR a search
direction and the other required arguments.